Split test helpers into separate file
Also made a couple of small changes to the API URL routes based on
patterns used in the last code challenge, which are informing the test
directory changes.
